Use `htmlStash` for pre-rendered block references
This change makes regdown use Python-Markdown’s `htmlStash` for the pre-rendered HTML content of `see(foo)` block references, instead of reparsing it with ElementTree and appending to the parent.

The `parent` in this case is an empty node, as the `see(foo)` is supposed to occur on a line byitself in the Markdown, so the `htmlStash` reference gets inserted as the parent’s text node.

This should fix an issue we’re seeing where we render HTML with an `<svg xmlns=“…”>` in it, and the namespace gets rendered out in the end as an ElementTree prefix of the `svg` node.